Moved all controller tests to "ControllerTests" package
